Implement —disable-assign-existing-users flag.
With this flag organization and application admin users are not able to assign existing users to an organization or application. Assigining an existing user requires the user to have access to the global username list to select the user to be assigned. This exposure is not always desired.

Global admin users are still able to assign existing users to organizations or application.