main.go: Make it possible to run gRPC on different port than REST/HTTP
As described in https://github.com/grpc/grpc-go/issues/555, it is not
possible to run gRPC without security when using net/http and serveHTTP.
This creates a problem if you want to run lora-app-server behind a proxy
which terminates TLS.

This problem is solved by adding three new options: GRPC_BIND,
GRPC_TLS_CERT and GRPC_TLS_KEY.

When GRPC_BIND is set, the gRPC service will bind to this ip:port
instead of HTTP_BIND which makes it possible to leave HTTP_TLS_CERT and
HTTP_TLS_KEY unset. GRPC_TLS_CERT and GRPC_TLS_KEY are optional and will
only be used if GRPC_BIND is set.