Add support for OpenID Connect authentication and registration.
This makes it possible to integrate ChirpStack Application Server with
for example Auth0.com. Using the registration_callback_url it is
possible to automatically provision new users in ChirpStack, create an
organization for the user etc.

In order to add support for OpenID Connect, the username of the user
object has changed into email. Existing users will have their email set
to their username and can continue to login as before. However, when
updating an user, the email field must be updated to a valid email!

Also the option to, as an organization admin, create new users and
assign them to organizations has removed. Assigning new users to an
organization is still possible by the user email address, but the
autocomplete has been removed to avoid privacy issues.

Closes #372.