Refactor MQTT integration topic configuration.
This aligns the MQTT configuration with the other components in that an
event and command topic template can be configured. When new events or
commands are introduced, this removes the need add additional
configuration.

This is fully backwards compatible with the old topic configuration.
When `_topic_template` configuration values are found in the
configuration file, these will override the `event_topic_template` and
`command_topic_template` values.