Refactor HTTP integration with single event endpoint url.
Instead of configuring a separate URL per event-type, this posts all
events to a single endpoint with an `eventType` url query argument
indicating the event-type.

This reduces the amount of configuration and does not require to create
new endpoints when new events are added.

This is fully backwards compatible as already configured endpoints for
specific events will override the new event endpoint url.