Remove Azure SB module and use HTTP for Azure Service Bus integration.
This makes the integration easier to maintain as it does not keep a
connection state with the Azure Service Bus endpoint. As the Azure
Service Bus does not deal with re-connect logic, this had to be
implemented within the integration.

Since the integration does no longer keep a connection state, it can be
made available through the web-interface as there is no downside when
many instances of this integration exist for different applications.