Implement ReEncryptDeviceQueueItems method to AS API.
This method is used by the ChirpStack Network Server when it needs to
update the frame-counter sequence of the queue-items, e.g. when the
downlink frame-counter was incremented by a mac-command only downlink.

Note: changing the frame-counter of a downlink requires re-encryption as
the frame-counter is one of the encryption parameters.