Update PostgreSQL integration uplink message storage.
This updates the rx_info JSON encoding so that it follows the configured
marshaler in the chirpstack-application-server.toml configuration. If
marshaler is set to 'protobuf', then the `json` marshaler (based on the
Protobuf messages) will be used, as the rx_info field must contain a
JSON value. For backwards compatibility, configure the `json_v3`
marshaler.

This also adds the tx_info field containing the TX related meta-data.
This field is also encoded using the configured marshaler.