Refactor multicast groups.
This moves the mulitcast feature under applications, to enforce that all
devices within the multicast group belong to the same application as the
multicast group.

Before you upgrade, please validate that if you have any existing
multicast-groups that:

1. These are not empty (they contain one or multiple devices)
2. All devices within the multicast group belong to the same
application

If these conditions are met, ChirpStack Application Server will
automatically migrate your multicast group(s) to the same application as
your devices. In any other case, the migration will throw an
error asking you to either remove the empty multicast group(s) or to
re-assign the devices so that the above criteria are met.