Use Redis Stream ID for de-duplication of device events.
The Redis Stream is incremented for each event (by Redis). Using the
PublishedAt timestamp has the risk of race-conditions as the timestamp
is set on constructing the message, not on publishing it. Therefore it
is possible that the latest message has a timestamp before the message
that was published before, in which case the web-interface would ignore
it.